% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk8dx-package.R
\docType{package}
\name{mk8dx-package}
\alias{mk8dx}
\alias{mk8dx-package}
\title{mk8dx: Access Mario Kart Speedrun Data}
\description{
Conveniently convert your Mario Kart lss speedrun files into table data.
}
\author{
\strong{Maintainer}: Sarah Johnson \email{sarahviamemphis@gmail.com}

}
\keyword{internal}
