% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss.R
\name{mk_lss}
\alias{mk_lss}
\title{Convert lss data into a table}
\usage{
mk_lss(filepath)
}
\arguments{
\item{filepath}{A path to an lss file.}
}
\value{
A tibble of the data for the lss file.
}
\description{
Convert lss data into a table
}
