% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss_folder.R
\name{mk_lss_folder}
\alias{mk_lss_folder}
\title{Get all lss files in a folder}
\usage{
mk_lss_folder(folderpath)
}
\arguments{
\item{folderpath}{A folder containing lss files}
}
\value{
A table of all lss data in the folder
}
\description{
Get all lss files in a folder
}
