% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_repair.R
\name{mk_repair}
\alias{mk_repair}
\title{Repair MK8DX columns}
\usage{
mk_repair(df)
}
\arguments{
\item{df}{A table of MK8DX speedrunning data}
}
\value{
Repaired table
}
\description{
Repair MK8DX columns
}
