% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_repair.R
\name{mk_repair_cup_cats}
\alias{mk_repair_cup_cats}
\title{Repair MK8DX cups in category column}
\usage{
mk_repair_cup_cats(df, category, individual_cup)
}
\arguments{
\item{df}{A table of MK8DX speedrunning data.}

\item{category}{The category with incorrect data}

\item{individual_cup}{The column for cup names}
}
\value{
Table with corrected cups and categories columns.
}
\description{
Repair MK8DX cups in category column
}
