% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_repair.R
\name{mk_repair_fill_vars}
\alias{mk_repair_fill_vars}
\title{Repair missing variables based on categories}
\usage{
mk_repair_fill_vars(df, category, individual_cup)
}
\arguments{
\item{df}{table of MK8DX speedrunning data.}

\item{category}{The categories to group by.}

\item{individual_cup}{The cups to group by.}
}
\value{
table with filled in values
}
\description{
Repair missing variables based on categories
}
