% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_repair.R
\name{mk_repair_outdated}
\alias{mk_repair_outdated}
\title{Repair outdated data}
\usage{
mk_repair_outdated(df, attempt_count, best_segment_time)
}
\arguments{
\item{df}{table of MK8DX speedrunning data.}

\item{attempt_count}{column to correct}

\item{best_segment_time}{column to correct}
}
\value{
table with corrected columns
}
\description{
Repair outdated data
}
