% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_track_names.R
\name{mk_track_names}
\alias{mk_track_names}
\title{Rename MK8DX track names}
\usage{
mk_track_names(
  df,
  style = "both",
  tag_position = "after",
  tag_surrounds = "brackets"
)
}
\arguments{
\item{df}{working df}

\item{style}{The track name as "full", "abbr", or "both"}

\item{tag_position}{Either "before" or "after" track name}

\item{tag_surrounds}{Either "parenthesis", "brackets", or "none"}
}
\value{
A dataframe with renamed track names
}
\description{
Rename MK8DX track names
}
