% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss.R
\name{mk_variables}
\alias{mk_variables}
\title{Get run variables from lss data.}
\usage{
mk_variables(lss)
}
\arguments{
\item{lss}{MK8DX XML data}
}
\value{
A tibble of the metadata variables for the lss file.
}
\description{
Get run variables from lss data.
}
