% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss.R
\name{mk_attempts}
\alias{mk_attempts}
\title{Run attempt history}
\usage{
mk_attempts(lss)
}
\arguments{
\item{lss}{MK8DX XML data}
}
\value{
A tibble of your run attempts for the lss file.
}
\description{
Run attempt history
}
