% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_attempts.R
\name{mk_attempts}
\alias{mk_attempts}
\title{Run attempt history}
\usage{
mk_attempts(filepath)
}
\arguments{
\item{filepath}{A filepath to an lss file.}
}
\value{
A tibble of your run attempt history for the lss file.
}
\description{
Run attempt history
}
