% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss.R
\name{mk_lss}
\alias{mk_lss}
\title{Convert a folder of lss files into a table.}
\usage{
mk_lss(folder)
}
\arguments{
\item{folder}{A folder path to your MK8DX lss files.}
}
\value{
A tibble of your speedrun data.
}
\description{
Convert a folder of lss files into a table.
}
