% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_segment_times.R
\name{mk_segment_times}
\alias{mk_segment_times}
\title{Get track times from an lss file}
\usage{
mk_segment_times(lss)
}
\arguments{
\item{lss}{An lss file read into R.}
}
\value{
A tibble of track times.
}
\description{
Get track times from an lss file
}
