% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_lss.R
\name{mk_segments}
\alias{mk_segments}
\title{Get segment names/id from an lss file.}
\usage{
mk_segments(lss)
}
\arguments{
\item{lss}{MK8DX XML data}
}
\value{
A tibble of tracks in the lss file.
}
\description{
Get segment names/id from an lss file.
}
