% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_variables.R
\name{mk_variables}
\alias{mk_variables}
\title{Get run variables from lss data.}
\usage{
mk_variables(lss)
}
\arguments{
\item{lss}{An lss file read into R.}
}
\value{
A tibble of the metadata variables for the lss file.
}
\description{
Get run variables from lss data.
}
